Add emulator.py with Plugboard class
emulator.py is the file in which all the actual emulation logic will be.
It has the Plugboard class in it, which represents the plugboard on an
Enigma machine. The plugboard performed a configurable substitution
cipher determined by the key. The operator could swap two letters, which
would be swapped before entering the rotors.

To swap two letters, call `swap()` with the two letters you wish to
swap. Alternatively, call `swap_all()` to swap a list of characters as
varargs. To get the encoding of a letter through the plugboard, call
`encode()` with your letter and the encoding will be returned. To reset
the plugboard, call `reset()`.