Add Enigma.encrypt(), Rotor.reverse_encrypt()
Enigma.encrypt() is the function that ties it all together: it encrypts
a character from start to finish, taking care of all side effects along
the way.

Rotor.reverse_encrypt() simulates running a signal through a rotor
backwards, as what happens after the signal passes through the
reflector. It's the exact reverse of Rotor.encrypt().