Add optional fourth rotor, thin reflectors/rotors
There were some Enigma models, such as the naval M4, which had space for
a fourth rotor. This fourth rotor did not turnover, adjustable only by
its ring position. It was fit into the same space as the reflector by
making both it and the reflector thin.

This is modeled through Enigma.__init__()'s optional `rotor4` param and
the new `thin` property of Rotors. If rotor4 is present, both it and
the reflector must have their `thin` property set to True.